refactor: remane modalReduser folder to modalSlice
