--- src/redux/modalReducer/index.ts
@@ -1,2 +1,0 @@
-export { default } from './modalReducer';
-export * from './modalReducer';
